# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AeroTwin API
  version: 1.0.0
paths:
...
  /api/v1/users:
    get:
      tags:
      - users
      summary: List Users
      description: List all users. Admin only.
      operationId: list_users_api_v1_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
                title: Response List Users Api V1 Users Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - users
      summary: Create User
      description: Create a new user. Admin only.
      operationId: create_user_api_v1_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/users/{user_id}:
    get:
      tags:
      - users
      summary: Get User
      description: Get a single user by ID. Admin only.
      operationId: get_user_api_v1_users__user_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - users
      summary: Update User
      description: Update a user. Admin only.
      operationId: update_user_api_v1_users__user_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - users
      summary: Delete User
      description: Delete a user. Admin only.
      operationId: delete_user_api_v1_users__user_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/simulation/run:
    post:
      tags:
      - simulation
      summary: Create Simulation Run
      description: 'Kick off a replay (historical mission, re-run through the ML pipeline

        at accelerated speed) or a what-if (synthetic scenario) simulation.

        Runs in the background; poll `GET /simulation/{id}` for the result.'
      operationId: create_simulation_run_api_v1_simulation_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimulationRunCreate'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/simulation/what-if/config:
    get:
      tags:
      - simulation
      summary: Get What If Config
      description: Slider metadata (min/max/step/unit) and window settings for the What-If UI.
      operationId: get_what_if_config_api_v1_simulation_what_if_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - OAuth2PasswordBearer: []
  /api/v1/simulation/what-if:
    post:
      tags:
      - simulation
      summary: Run What If Scenario
      description: 'Parameter-perturbation What-If: runs the engine''s recent REAL telemetry

        window, with the requested parameter change applied in memory to the

        last K readings, through the existing RUL/Fault/Bearing/Aux models and

        Health Fusion, and compares against the unmodified window.


        Strictly read-only: writes no telemetry, predictions, health scores or

        alerts, broadcasts nothing, and does not touch live model state.

        Synchronous (returns the finished result; `simulation_status` says how

        it ended).'
      operationId: run_what_if_scenario_api_v1_simulation_what_if_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WhatIfRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/simulation/{simulation_id}:
    get:
      tags:
      - simulation
      summary: Get Simulation Run
      operationId: get_simulation_run_api_v1_simulation__simulation_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: simulation_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Simulation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SimulationRunCreate:
      properties:
        engine_id:
          type: string
          format: uuid
          title: Engine Id
        mode:
          type: string
          enum:
          - replay
          - what_if
          title: Mode
        mission_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Mission Id
        environmental_profile:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Environmental Profile
      type: object
      required:
      - engine_id
      - mode
      title: SimulationRunCreate
...
    User:
      properties:
        email:
          type: string
          format: email
          title: Email
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
          default: true
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
      - email
      - role
      - id
      title: User
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        role:
          type: string
          title: Role
        is_active:
          type: boolean
          title: Is Active
          default: true
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - role
      - password
      title: UserCreate
    UserUpdate:
      properties:
        email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Email
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        password:
          anyOf:
          - type: string
          - type: 'null'
          title: Password
      type: object
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WhatIfRequest:
      properties:
        engine_id:
          type: string
          format: uuid
          title: Engine Id
        baseline_timestamp:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Baseline Timestamp
        parameters:
          additionalProperties: true
          type: object
          title: Parameters
      type: object
      required:
      - engine_id
      - parameters
      title: WhatIfRequest
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login